SimpleLink lpf3-7.20.01.10 update This commit updates to the SimpleLink F3 7.20.01.10 SDK release.

+++ source/ti/ble5stack_flash/osal/src/inc/osal_snv.h
@@ -77,7 +77,7 @@ extern "C"
  */
 
   typedef uint16 osalSnvId_t;
-  typedef uint8  osalSnvLen_t;
+  typedef uint16 osalSnvLen_t;
 
 /*********************************************************************
  * GLOBAL VARIABLES
